refactor: 💡 virtual scroller と layout を分離
