+++ src/components/UI/virtualScroller.ts
@@ -1,0 +1,91 @@
+import { derived, writable } from "svelte/store";
+
+export const useVirtualScroller = () => {
+  const virtualHeight = writable(0);
+  const setVirtualHeight = (value: number) => virtualHeight.set(value);
+
+  const scrollY = writable(0);
+  const containerHeight = writable(0);
+
+  const headerHeight = writable(0);
+
+  const contentsWidth = writable(0);
+
+  const contentsScrollY = derived<
+    [typeof scrollY, typeof headerHeight],
+    number
+  >([scrollY, headerHeight], ([$scrollY, $headerHeight], set) => {
+    set(Math.max(0, $scrollY - $headerHeight));
+  });
+
+  const header = (node: HTMLElement) => {
+    const resizeObserver = new ResizeObserver((entries) => {
+      const entry = entries[0];
+      if (!entry) {
+        return;
+      }
+      headerHeight.set(entry.contentRect.height);
+    });
+    resizeObserver.observe(node);
+
+    return {
+      destroy() {
+        resizeObserver.disconnect();
+      },
+    };
+  };
+  const contents = (node: HTMLElement) => {
+    const resizeObserver = new ResizeObserver((entries) => {
+      const entry = entries[0];
+      if (!entry) {
+        return;
+      }
+      contentsWidth.set(entry.contentRect.width);
+    });
+    resizeObserver.observe(node);
+
+    return {
+      destroy() {
+        resizeObserver.disconnect();
+      },
+    };
+  };
+
+  const container = (node: HTMLElement) => {
+    const onScroll = (e: Event) => {
+      const target = e.target;
+      if (!(target instanceof HTMLElement)) {
+        return;
+      }
+      scrollY.set(target.scrollTop);
+    };
+    node.addEventListener("scroll", onScroll);
+
+    const resizeObserver = new ResizeObserver((entries) => {
+      const entry = entries[0];
+      if (!entry) {
+        return;
+      }
+      containerHeight.set(entry.contentRect.height);
+    });
+    resizeObserver.observe(node);
+
+    return {
+      destroy() {
+        node.removeEventListener("scroll", onScroll);
+        resizeObserver.disconnect();
+      },
+    };
+  };
+
+  return {
+    container,
+    header,
+    contents,
+    virtualHeight,
+    setVirtualHeight,
+    contentsWidth,
+    contentsScrollY,
+    containerHeight,
+  };
+};
